Refactor file creation and add error handling when directory cannot be created
